@@ -6,20 +6,20 @@ namespace BooksLib.Models
     {
         public int BookId { get; set; }
 
-        private string _title;
-        public string Title
+        private string? _title;
+        public string? Title
         {
             get => _title;
             set => SetProperty(ref _title, value);
         }
 
-        private string _publisher;
-        public string Publisher
+        private string? _publisher;
+        public string? Publisher
         {
             get => _publisher;
             set => SetProperty(ref _publisher, value);
         }
 
-        public override string ToString() => Title;
+        public override string ToString() => Title ?? string.Empty;
     }
 }
