@@ -7,15 +7,15 @@ namespace BooksLib.Repositories
 {
     public class BooksSampleRepository : IBooksRepository
     {
-        private List<Book> _books;
+        private readonly List<Book> _books;
         public BooksSampleRepository()
         {
-            InitSampleBooks();
+            _books = InitSampleBooks();
         }
 
-        private void InitSampleBooks()
+        private List<Book> InitSampleBooks()
         {
-            _books = new List<Book>()
+            return new List<Book>()
             {
                 new Book { BookId = 1, Title = "Professional C# 6 and .NET Core 1.0", Publisher = "Wrox Press" },
                 new Book { BookId = 2, Title = "Professional C# 5.0 and .NET 4.5.1", Publisher = "Wrox Press" },
@@ -25,7 +25,7 @@ namespace BooksLib.Repositories
 
         public Task<bool> DeleteAsync(int id)
         {
-            Book bookToDelete = _books.Find(b => b.BookId == id);
+            Book? bookToDelete = _books.Find(b => b.BookId == id);
             if (bookToDelete != null)
             {
                 return Task.FromResult<bool>(_books.Remove(bookToDelete));
@@ -33,19 +33,19 @@ namespace BooksLib.Repositories
             return Task.FromResult<bool>(false);
         }
 
-        public Task<Book> GetItemAsync(int id) =>
+        public Task<Book?> GetItemAsync(int id) =>
             Task.FromResult(_books.Find(b => b.BookId == id));
 
         public Task<IEnumerable<Book>> GetItemsAsync() =>
             Task.FromResult<IEnumerable<Book>>(_books);
 
-        public Task<Book> UpdateAsync(Book item)
+        public Task<Book?> UpdateAsync(Book item)
         {
-            Book bookToUpdate = _books.Find(b => b.BookId == item.BookId);
-            if (bookToUpdate == null) return null;
+            Book? bookToUpdate = _books.Find(b => b.BookId == item.BookId);
+            if (bookToUpdate == null) return Task.FromResult<Book?>(null);
             int ix = _books.IndexOf(bookToUpdate);
             _books[ix] = item;
-            return Task.FromResult(_books[ix]);
+            return Task.FromResult<Book?>(_books[ix]);
         }
 
         public Task<Book> AddAsync(Book item)
