@@ -7,7 +7,7 @@ namespace BooksLib.Services
     public interface IBooksService
     {
         Task<Book> AddOrUpdateBookAsync(Book book);
-        Book GetBook(int bookId);
+        Book? GetBook(int bookId);
         Task LoadBooksAsync();
 
         IEnumerable<Book> Books { get; }
