@@ -4,9 +4,9 @@ using System.Runtime.CompilerServices;
 
 namespace Framework
 {
-    public class BindableObject : INotifyPropertyChanged
+    public abstract class BindableObject : INotifyPropertyChanged
     {
-        public event PropertyChangedEventHandler PropertyChanged;
+        public event PropertyChangedEventHandler? PropertyChanged;
 
         protected virtual void OnPropertyChanged([CallerMemberName] string propertyName = null) =>
             PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
