@@ -5,22 +5,19 @@ namespace UnitTestingSamples
 {
     public class Formula1
     {
-        private IChampionsLoader loader;
-        public Formula1(IChampionsLoader loader)
-        {
-            this.loader = loader;
-        }
+        private IChampionsLoader _loader;
+        public Formula1(IChampionsLoader loader) => _loader = loader;
 
         public XElement ChampionsByCountry(string country)
         {
-            var q = from r in loader.LoadChampions().Elements("Racer")
-                    where r.Element("Country").Value == country
-                    orderby int.Parse(r.Element("Wins").Value) descending
+            var q = from r in _loader.LoadChampions().Elements("Racer")
+                    where r.Element("Country")?.Value == country
+                    orderby int.Parse(r.Element("Wins")?.Value ?? "0") descending
                     select new XElement("Racer",
-                      new XAttribute("Name", r.Element("Firstname").Value + " " +
-                        r.Element("Lastname").Value),
-                      new XAttribute("Country", r.Element("Country").Value),
-                      new XAttribute("Wins", r.Element("Wins").Value));
+                      new XAttribute("Name", r.Element("Firstname")?.Value + " " +
+                        r.Element("Lastname")?.Value),
+                      new XAttribute("Country", r.Element("Country")?.Value ?? "unknown"),
+                      new XAttribute("Wins", r.Element("Wins")?.Value ?? "0"));
             return new XElement("Racers", q.ToArray());
         }
 
